A follow-up to #1589
Adresses #1690

A number of optimisations around bindings. Apparently there is not much sense in adding table indexes to binding tables, as index read locks entire table. Indexes could slightly improve performance when there are many bindings and queues, but will require a breaking change in the schema.

The changes in this branch provide performance improvement comparable to introduction of indexes without reworking the schema:

- Replaced `delete_object` with `delete` as bindings are value-objects and there are no bag tables
- In some lookups during cleanup `match_object` is replaced with `dirty_match_object` to not lock the entire bindings table. This change was reverted in d0423f958599cba5148864997aa8446c873dad00 because it was creating inconsistencies with `dirty_delete`. Current branch does not contain dirty deletes. Still requires more testing.
- Default bindings are removed. The default exchange does direct lookup of queues without actually loading the bindings and default bindings are not unbindable and only displayed in `list_bindings`. They are not used, but slow down queue deletion and creation.

Benchmark data will follow.