@hairyhum 2018-09-28
Exclusive queue cleanup benchmark:

Time to cleanup queues on connection:

| 100K queues | No explicit bindings | 1 binding | 2 bindings |
|--------------|---------------------|----------|------------|
| Before            | 57.7 sec                     | 77.9 sec  | 107.1 sec   |
| After               | 20.9 sec                    | 25.7 sec  | 30.5 sec   |