@hairyhum 2018-09-28
This can be backported to 3.7 if accepted.