@hairyhum 2018-10-01
The interesting thing about default bindings and min-master locator is that it's counting queue masters looking at bindings, but all queues were bound. Originally it was counting bindings and selecting a node which min bindings to the queues on the node. Then it was changed in https://github.com/rabbitmq/rabbitmq-server/commit/1d413eac80c2adb6844b82c8ce57e21a04a55b04 to count multiple bindings for the same queue as one. The assumption was that we should not account for unbound queues. But  any queue had at least 1 binding (default).
I'm going to make it count queues, because it will be the same behaviour.