@hairyhum 2018-10-02
So default bindings are shown in the management API. I've missed that somehow. I still don't think we should store them in the database. Maybe we could fake them for the management API
I'm going to create a separate PR with default bindings removed.