@hairyhum 2018-10-02
Moved default bindings removal to a separate PR since most of the failures are about that change. It might be even for 3.8.