@michaelklishin 2018-10-02
@hairyhum we can append them in HTTP API handlers and `rabbitmqctl list_bindings` in CLI.