@michaelklishin 2018-10-02
…or functions that list bindings, that'd cover more cases with a single change?