@hairyhum 2018-10-03
Merged to 3.7